+++ ouds_theme_contract/lib/theme/tokens/components/ouds_components_tokens.dart
@@ -12,17 +12,22 @@
 
 import 'package:ouds_theme_contract/ouds_tokens_provider.dart';
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_button_tokens.dart';
+import 'package:ouds_theme_contract/theme/tokens/components/ouds_radioButton_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_skeleton_tokens.dart';
 
 class OudsComponentsTokens {
   final OudsButtonTokens button;
   final OudsSkeletonTokens skeleton;
+  final OudsRadioButtonTokens radioButton;
 
   OudsComponentsTokens({
     required OudsProvidersTokens providersTokens,
     OudsButtonTokens? button,
     OudsSkeletonTokens? skeleton,
+    OudsRadioButtonTokens? radioButton,
   })  : button = button ?? OudsButtonTokens(providersTokens: providersTokens),
         skeleton =
-            skeleton ?? OudsSkeletonTokens(providersTokens: providersTokens);
+            skeleton ?? OudsSkeletonTokens(providersTokens: providersTokens),
+        radioButton = radioButton ?? OudsRadioButtonTokens(providersTokens: providersTokens);
+
 }